unify users.controller.test and users.api.test
